pyproject.toml
[project]
name = "myjournalisticcrew"
version = "0.1.0"
description = "international news agency"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "annotated-types>=0.7.0",
    "crewai[tools]>=0.108.0,<1.0.0",
    "neo4j>=5.28.1",
    "selenium>=4.30.0",
    "webdriver-manager>=4.0.2",
    "python-dotenv>=1.0.0",
    "pillow>=11.1.0",
    "pygooglenews>=0.1.3",
    "undetected-chromedriver>=3.5.5",
    "selenium-wire-2>=0.2.1",
    "blinker>=1.9.0",
    "appdirs>=1.4.4",
    "auth0-python>=4.9.0",
    "boto3>=1.37.33",
    "aiohttp>=3.11.16",
    "beautifulsoup4>=4.13.3",
    "vercel-blob>=0.3.0",
]

[project.scripts]
myjournalisticcrew = "my_journalistic_crew.main:run"
run_crew = "my_journalistic_crew.main:run"
train = "my_journalistic_crew.main:train"
replay = "my_journalistic_crew.main:replay"
test = "my_journalistic_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.hatch.build.targets.wheel]
packages = ["src/my_journalistic_crew"]
